one more migration (and hopefully the last today)
i am getting a little tired of copying the logs to
refs/notes/vercel/deploy

but basically, vercel has to actually BE in the same directory
for the functions to be copied.  contrary to what GPT made me
believe, it does NOT do such smart thing as perform a static
analysis of the code and figure out all the modules referenced
by the entrypoint script, and the modules referenced by those,
then by those again, so on and so forth.... no.  they MUST be
in the same directory.

how do i know well.... people usually don't trouble themselves
with extra logic, or purposefully burden people with the lack
of that.  judging by the way some of the previous deployments
worked, i assume this must be not very much different....