add even more exhaustive Mac OS + Windows ignores
i guess i don't blame people as much.... researching
this list alone and trying to get a grasp of why apple
made these choices is a pure headache @_@

like i tried to get a hold of why Mac OS named this peculiar
file "Icon\r" - like i know .DS_Store = Desktop Services Store,
.Spotlight is the Cmd+Space spotlight search cache, all of
these had (somewhat) logical names and, more importantly, start
with a dot like the rest of Unix descendents do.  But at SOME
point of time, Mac OS wasn't Unix-based as it has been since OS X
(with an XNU (unrelated to GNU) kernel and a FreeBSD userspace)
and did things its own way such as using CR as the line delimiter.

The relevance of this information is that "Icon\r" ends with
** that very same CR. **  So SOMEONE (namely GPT) says that
"Icon\r" was named that way because it was (classic) Mac OS's
way of hiding a file (before adopting the Unix convention of
using dot-files that is).  But, being unable to re-produce the
result on my MacBook Finder ("Icon\r" always shows up as "Icon?"
which isn't even a special character for "\r"; "Icon\n" shows
the same), i cannot say i am ready to believe that was the case.

the more compelling reason was that "Icon\r" was only a link to
something else called Resource Fork which is FAR past the scope
of this discussion.  either way, the links are there.... maybe
they could be of use, maybe not?  or maybe i forgot if anyone on
the team even uses a MacBook.  eh. :v

(I think I can name at least one person that uses Mac in my head
so all of this shouldn't be a complete waste of time.)