still doesn't work.
so apparently it was calling the terminal WSGI because
none of the middleware could handle it.  weird.  on the
other hand, vite seems to be strict about the Host header
matching 'localhost' literally so i'm gonna have to fix
the URL.  still doesn't work though because it either
falls through the terminal WSGI or ends up unhandle-able
by the WSGI in the middle.  (i wonder if the middleware
is stripping the "/api/" prefix?  Let me check the docs
for DispatcherMiddleware then.)