Sample search bar from GPT
so i asked gpt to write me a console log because i didn't
want to scare people off by telling them to open devtools,
but ah.... it's hard.  it's fricking hard. :,)  because to
export OTHER subroutines from a svelte component you need
to do THIS crazy thing:
<https://stackoverflow.com/q/58287729/19411800>

... which i obviously do not have the mental capacity to
understand.  so here is that done differently anyways.  it's
cumbersume: the state is stored in debug/conctl.ts, while
the component is stored in debug/Console.svelte.  but it's
the only way i can do something as crazy as (ab)using strbuf
as a global state.... so i guess that's that :/

oh and the backend is absolutely dumb uh i guess they won't
have a chance to actually connect to any SQLite database
(but then again they can just json or something for that...
I'LL DO THAT.  these people just won't do anything on their
own.....)